va-qc-app-old/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
python-dotenv==1.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-cloud-storage==2.14.0
pydantic==1.10.13
bcrypt==4.1.2
itsdangerous==2.1.2
starlette==0.35.0
